--- my-app/src/test/java/com/mycompany/app/TicTacToeCellTest.java
@@ -1,41 +1,0 @@
-package com.mycompany.app;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-
-public class TicTacToeCellTest {
-
-    @BeforeAll
-    static void headless() {
-        System.setProperty("java.awt.headless", "true");
-    }
-
-    @Test
-    void constructorStoresCoordinates() {
-        TicTacToeCell cell = new TicTacToeCell(4, 1, 2);
-        assertEquals(4, cell.getNum());
-        assertEquals(1, cell.getCol());
-        assertEquals(2, cell.getRow());
-        assertEquals(' ', cell.getMarker());
-    }
-
-    @Test
-    void setMarkerUpdatesStateAndDisablesCell() {
-        TicTacToeCell cell = new TicTacToeCell(0, 0, 0);
-        cell.setMarker("X");
-        assertEquals('X', cell.getMarker());
-        assertEquals("X", cell.getText());
-        assertFalse(cell.isEnabled());
-    }
-
-    @Test
-    void setMarkerOUpdatesText() {
-        TicTacToeCell cell = new TicTacToeCell(8, 2, 2);
-        cell.setMarker("O");
-        assertEquals('O', cell.getMarker());
-        assertEquals("O", cell.getText());
-    }
-}
