--- my-app/src/test/java/com/mycompany/app/TicTacToePanelTest.java
@@ -1,24 +1,0 @@
-package com.mycompany.app;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import java.awt.GridLayout;
-
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-
-public class TicTacToePanelTest {
-
-    @BeforeAll
-    static void headless() {
-        System.setProperty("java.awt.headless", "true");
-    }
-
-    @Test
-    void panelConstructorCreatesNineCells() {
-        TicTacToePanel panel = new TicTacToePanel(new GridLayout(3, 3));
-        assertNotNull(panel);
-        assertEquals(9, panel.getComponentCount());
-    }
-}
