+++ report/README.md
@@ -1,2 +1,2 @@
 Здесь сохраняются скриншоты отчетов о покрытии
-![img_1.png](img_1.png)
+![img_3.png](img_3.png)~~~~
